requirement.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-celery-beat==2.7.0
django-extensions==3.2.3
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.10
kombu==5.4.2
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
